groq-builder: feature(compat): renamed to `compat`
