+++ packages/groq-builder/src/compat.ts
@@ -1,4 +1,6 @@
 import { validate } from "./validation/validate";
-import { createGroqBuilder } from "./index";
+import { createGroqBuilder, makeSafeQueryRunner } from "./index";
 
 export const q = Object.assign(createGroqBuilder<any>(), validate);
+
+export { makeSafeQueryRunner };
